requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.20
pdfplumber==0.10.2
PyPDF2==3.0.1
Pillow==10.1.0
pdf2image==1.16.3
pytesseract==0.3.10
sentence-transformers>=2.2.0
chromadb>=0.4.0
langchain>=0.0.200
python-docx==1.1.2